+++ src/ts/components/core/NavLink.tsx
@@ -6,9 +6,11 @@ import {
 import { BoxProps } from "props/box";
 import { DashBaseProps, PersistenceProps } from "props/dash";
 import { StylesApiProps } from "props/styles";
-import React, { MouseEvent } from "react";
+import React, { MouseEvent, useState, useEffect } from "react";
 import { TargetProps, onClick } from "../../utils/anchor";
+import { History } from '@plotly/dash-component-plugins';
 import { setPersistence, getLoadingState } from "../../utils/dash3";
+
 
 interface Props
     extends BoxProps,
@@ -23,8 +25,12 @@ interface Props
     leftSection?: React.ReactNode;
     /** Section displayed on the right side of the label */
     rightSection?: React.ReactNode;
-    /** Determines whether the link should have active styles, `false` by default */
-    active?: boolean;
+    /**
+     * Controls whether the link is styled as active (default: `false`).
+     * - `exact`: Active if `pathname` matches `href` exactly.
+     * - `partial`: Active if `pathname` starts with `href` (for subpages).
+     */
+    active?: boolean | 'exact' | 'partial';
     /** Key of `theme.colors` of any valid CSS color to control active styles, `theme.primaryColor` by default */
     color?: MantineColor;
     /** href */
@@ -64,8 +70,28 @@ const NavLink = ({
     persistence_type,
     setProps,
     loading_state,
+    active = false,
     ...others
 }: Props) => {
+    const [linkActive, setLinkActive] = useState(false);
+
+    const pathnameToActive = pathname => {
+        setLinkActive(
+          active === true ||
+            (active === 'exact' && pathname === href) ||
+            (active === 'partial' && (pathname.startsWith(href + '/') || pathname === href))
+        );
+    };
+
+    useEffect(() => {
+        pathnameToActive(window.location.pathname);
+
+        if (typeof active === 'string') {
+          History.onChange(() => {
+            pathnameToActive(window.location.pathname);
+          });
+        }
+    }, [active]);
 
     const onChange = (state: boolean) => {
         setProps({ opened: state });
@@ -91,6 +117,7 @@ const NavLink = ({
                 target={target}
                 onChange={onChange}
                 disabled={disabled}
+                active={linkActive}
                 {...others}
             >
                 {children}
@@ -102,6 +129,7 @@ const NavLink = ({
                 disabled={disabled}
                 onChange={onChange}
                 onClick={increment}
+                active={linkActive}
                 {...others}
             >
                 {children}